CGBA-40 Push message parsing and validation logic down into BalanceRequestService
