+++ test/services/BalanceRequestCacheServiceSpec.scala
@@ -25,17 +25,12 @@ import config.AppConfig
 import connectors.FakeEisRouterConnector
 import metrics.FakeMetrics
 import models.BalanceRequestFunctionalError
-import models.BalanceRequestResponse
 import models.BalanceRequestSuccess
 import models.MessageType
 import models.PendingBalanceRequest
-import models.SchemaValidationError
-import models.errors.BadRequestError
-import models.errors.BalanceRequestError
 import models.errors.FunctionalError
-import models.errors.NotFoundError
+import models.errors.InternalServiceError
 import models.errors.UpstreamTimeoutError
-import models.errors.XmlValidationError
 import models.request.BalanceRequest
 import models.values._
 import org.scalatest.flatspec.AsyncFlatSpec
@@ -52,8 +47,6 @@ import java.util.UUID
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration._
 import scala.util.Random
-import scala.xml.Elem
-import scala.xml.NodeSeq
 import scala.xml.Utility
 
 class BalanceRequestCacheServiceSpec extends AsyncFlatSpec with Matchers {
@@ -88,6 +81,7 @@ class BalanceRequestCacheServiceSpec extends AsyncFlatSpec with Matchers {
       ),
       new XmlFormattingServiceImpl,
       validator,
+      parser,
       FakeEisRouterConnector(sendMessageResponse),
       appConfig,
       Clock.systemUTC()
@@ -96,9 +90,7 @@ class BalanceRequestCacheServiceSpec extends AsyncFlatSpec with Matchers {
     new BalanceRequestCacheServiceImpl(
       appConfig,
       new FakeMetrics,
-      balanceRequestService,
-      validator,
-      parser
+      balanceRequestService
     )
   }
 
@@ -334,7 +326,7 @@ class BalanceRequestCacheServiceSpec extends AsyncFlatSpec with Matchers {
     assertion.unsafeToFuture()
   }
 
-  "BalanceRequestCacheService.updateBalance" should "return balance request success response when everything is successful" in {
+  "BalanceRequestCacheService.updateBalance" should "return no errors when all goes well" in {
     val uuid        = UUID.fromString("22b9899e-24ee-48e6-a189-97d1f45391c4")
     val balanceId   = BalanceId(uuid)
     val enrolmentId = EnrolmentId("12345678ABC")
@@ -394,7 +386,7 @@ class BalanceRequestCacheServiceSpec extends AsyncFlatSpec with Matchers {
     )
 
     cacheService
-      .updateBalanceRequest(
+      .updateBalance(
         balanceId.messageIdentifier,
         MessageType.ResponseQueryOnGuarantees,
         validResponseXml
@@ -405,8 +397,9 @@ class BalanceRequestCacheServiceSpec extends AsyncFlatSpec with Matchers {
       .unsafeToFuture()
   }
 
-  it should "return XML validation error when the XML does not pass schema validation" in {
-    val responseMissingMesTypXml = Utility
+  it should "return internal service error if the underlying balance request is not updated with the response" in {
+
+    val validResponseXml = Utility
       .trim(
         <CD037A>
           <SynIdeMES1>UNOC</SynIdeMES1>
@@ -417,6 +410,7 @@ class BalanceRequestCacheServiceSpec extends AsyncFlatSpec with Matchers {
           <TimOfPreMES10>1505</TimOfPreMES10>
           <IntConRefMES11>{Random.alphanumeric.take(14).mkString}</IntConRefMES11>
           <MesIdeMES19>{Random.alphanumeric.take(14).mkString}</MesIdeMES19>
+          <MesTypMES20>GB037A</MesTypMES20>
           <TRAPRIRC1>
             <TINRC159>GB12345678900</TINRC159>
           </TRAPRIRC1>
@@ -443,159 +437,17 @@ class BalanceRequestCacheServiceSpec extends AsyncFlatSpec with Matchers {
       .toString
 
     val cacheService = service(
-      updateBalanceRequestResponse = IO.none
+      updateBalanceRequestResponse = IO.some(pendingBalanceRequest)
     )
 
     cacheService
-      .updateBalanceRequest(
+      .updateBalance(
         balanceId.messageIdentifier,
         MessageType.ResponseQueryOnGuarantees,
-        responseMissingMesTypXml
+        validResponseXml.toString
       )
       .map {
-        _ shouldBe Left(
-          XmlValidationError(
-            MessageType.ResponseQueryOnGuarantees,
-            NonEmptyList.one(
-              SchemaValidationError(
-                1,
-                332,
-                "cvc-complex-type.2.4.a: Invalid content was found starting with element 'TRAPRIRC1'. One of '{MesTypMES20}' is expected."
-              )
-            )
-          )
-        )
-      }
-      .unsafeToFuture()
-  }
-
-  it should "return bad request error if XML with missing data somehow passes schema validation" in {
-    val responseMissingDataXml =
-      <CD037A>
-        <SynIdeMES1>UNOC</SynIdeMES1>
-        <SynVerNumMES2>3</SynVerNumMES2>
-        <MesSenMES3>NTA.GB</MesSenMES3>
-        <MesRecMES6>MDTP-GUA-00000000000000000000001-01</MesRecMES6>
-        <DatOfPreMES9>20210806</DatOfPreMES9>
-        <TimOfPreMES10>1505</TimOfPreMES10>
-        <IntConRefMES11>{Random.alphanumeric.take(14).mkString}</IntConRefMES11>
-        <MesIdeMES19>{Random.alphanumeric.take(14).mkString}</MesIdeMES19>
-        <MesTypMES20>GB037A</MesTypMES20>
-        <TRAPRIRC1>
-          <TINRC159>GB12345678900</TINRC159>
-        </TRAPRIRC1>
-        <CUSTOFFGUARNT>
-          <RefNumRNT1>GB000001</RefNumRNT1>
-        </CUSTOFFGUARNT>
-        <GUAREF2>
-          <GuaRefNumGRNREF21>21GB3300BE0001067A001017</GuaRefNumGRNREF21>
-          <AccDatREF24>20210114</AccDatREF24>
-          <GuaTypREF22>4</GuaTypREF22>
-          <GuaMonCodREF23>1</GuaMonCodREF23>
-          <GUAQUE>
-            <QueIdeQUE1>2</QueIdeQUE1>
-          </GUAQUE>
-          <EXPEXP>
-            <ExpEXP1>2751.95</ExpEXP1>
-            <ExpCouEXP2>2448</ExpCouEXP2>
-            <CurEXP4>GBP</CurEXP4>
-          </EXPEXP>
-        </GUAREF2>
-      </CD037A>
-
-    val dummyXmlValidator = new XmlValidationService {
-      override def validate(
-        messageType: MessageType,
-        xml: String
-      ): Either[NonEmptyList[SchemaValidationError], Elem] =
-        Right(responseMissingDataXml)
-    }
-
-    val cacheService = service(
-      updateBalanceRequestResponse = IO.none,
-      validator = dummyXmlValidator
-    )
-
-    cacheService
-      .updateBalanceRequest(
-        balanceId.messageIdentifier,
-        MessageType.ResponseQueryOnGuarantees,
-        responseMissingDataXml.toString
-      )
-      .map {
-        _ shouldBe Left(BadRequestError("Unable to parse required values from IE037 message"))
-      }
-      .unsafeToFuture()
-  }
-
-  it should "return not found error if the underlying balance request can't be found" in {
-    val responseMissingDataXml =
-      <CD037A>
-        <SynIdeMES1>UNOC</SynIdeMES1>
-        <SynVerNumMES2>3</SynVerNumMES2>
-        <MesSenMES3>NTA.GB</MesSenMES3>
-        <MesRecMES6>MDTP-GUA-00000000000000000000001-01</MesRecMES6>
-        <DatOfPreMES9>20210806</DatOfPreMES9>
-        <TimOfPreMES10>1505</TimOfPreMES10>
-        <IntConRefMES11>{Random.alphanumeric.take(14).mkString}</IntConRefMES11>
-        <MesIdeMES19>{Random.alphanumeric.take(14).mkString}</MesIdeMES19>
-        <MesTypMES20>GB037A</MesTypMES20>
-        <TRAPRIRC1>
-          <TINRC159>GB12345678900</TINRC159>
-        </TRAPRIRC1>
-        <CUSTOFFGUARNT>
-          <RefNumRNT1>GB000001</RefNumRNT1>
-        </CUSTOFFGUARNT>
-        <GUAREF2>
-          <GuaRefNumGRNREF21>21GB3300BE0001067A001017</GuaRefNumGRNREF21>
-          <AccDatREF24>20210114</AccDatREF24>
-          <GuaTypREF22>4</GuaTypREF22>
-          <GuaMonCodREF23>1</GuaMonCodREF23>
-          <GUAQUE>
-            <QueIdeQUE1>2</QueIdeQUE1>
-          </GUAQUE>
-          <EXPEXP>
-            <ExpEXP1>2751.95</ExpEXP1>
-            <ExpCouEXP2>2448</ExpCouEXP2>
-            <CurEXP4>GBP</CurEXP4>
-          </EXPEXP>
-        </GUAREF2>
-      </CD037A>
-
-    val dummyXmlValidator = new XmlValidationService {
-      override def validate(
-        messageType: MessageType,
-        xml: String
-      ): Either[NonEmptyList[SchemaValidationError], Elem] =
-        Right(responseMissingDataXml)
-    }
-
-    val dummyXmlParser = new XmlParsingService {
-      override def parseResponseMessage(
-        messageType: MessageType,
-        message: NodeSeq
-      ): Either[BalanceRequestError, BalanceRequestResponse] =
-        Right(balanceRequestSuccess)
-    }
-
-    val cacheService = service(
-      updateBalanceRequestResponse = IO.none,
-      validator = dummyXmlValidator,
-      parser = dummyXmlParser
-    )
-
-    cacheService
-      .updateBalanceRequest(
-        balanceId.messageIdentifier,
-        MessageType.ResponseQueryOnGuarantees,
-        responseMissingDataXml.toString
-      )
-      .map {
-        _ shouldBe Left(
-          NotFoundError(
-            "The balance request with message identifier MDTP-GUA-22b9899e24ee48e6a18997d1 was not found"
-          )
-        )
+        _ shouldBe Left(InternalServiceError())
       }
       .unsafeToFuture()
   }
