+++ test/services/BalanceRequestServiceSpec.scala
@@ -16,15 +16,23 @@
 
 package services
 
+import cats.data.NonEmptyList
 import cats.effect.IO
 import cats.effect.unsafe.implicits.global
 import config.AppConfig
 import connectors.FakeEisRouterConnector
+import models.BalanceRequestResponse
 import models.BalanceRequestSuccess
+import models.MessageType
 import models.PendingBalanceRequest
+import models.SchemaValidationError
+import models.errors.BadRequestError
+import models.errors.BalanceRequestError
 import models.errors.InternalServiceError
+import models.errors.NotFoundError
 import models.errors.SelfCheckError
 import models.errors.UpstreamServiceError
+import models.errors.XmlValidationError
 import models.request.BalanceRequest
 import models.values._
 import org.scalatest.flatspec.AsyncFlatSpec
@@ -38,7 +46,10 @@ import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
 import java.time.Clock
 import java.time.Instant
 import java.util.UUID
+import scala.util.Random
 import scala.xml.Elem
+import scala.xml.NodeSeq
+import scala.xml.Utility
 
 class BalanceRequestServiceSpec extends AsyncFlatSpec with Matchers {
 
@@ -55,6 +66,8 @@ class BalanceRequestServiceSpec extends AsyncFlatSpec with Matchers {
     updateBalanceRequestResponse: IO[Option[PendingBalanceRequest]] = IO.stub,
     sendMessageResponse: IO[Either[UpstreamErrorResponse, Unit]] = IO.stub,
     formatter: XmlFormattingService = new XmlFormattingServiceImpl,
+    validator: XmlValidationService = new XmlValidationServiceImpl,
+    parser: XmlParsingService = new XmlParsingServiceImpl,
     appConfig: AppConfig = mkAppConfig(Configuration())
   ) = {
     val repository = FakeBalanceRequestRepository(
@@ -65,7 +78,8 @@ class BalanceRequestServiceSpec extends AsyncFlatSpec with Matchers {
     new BalanceRequestService(
       repository,
       formatter,
-      new XmlValidationServiceImpl,
+      validator,
+      parser,
       FakeEisRouterConnector(sendMessageResponse),
       appConfig,
       Clock.systemUTC()
@@ -272,29 +286,255 @@ class BalanceRequestServiceSpec extends AsyncFlatSpec with Matchers {
       .unsafeToFuture()
   }
 
-  "BalanceRequestService.updateBalanceRequest" should "delegate to repository" in {
+  "BalanceService.updateBalanceRequest" should "return updated balance request when everything is successful" in {
     val uuid        = UUID.fromString("22b9899e-24ee-48e6-a189-97d1f45391c4")
     val balanceId   = BalanceId(uuid)
     val enrolmentId = EnrolmentId("12345678ABC")
 
     val balanceRequestSuccess =
-      BalanceRequestSuccess(BigDecimal("12345678.90"), CurrencyCode("GBP"))
+      BalanceRequestSuccess(BigDecimal("1212211848.45"), CurrencyCode("GBP"))
 
     val updatedBalanceRequest = PendingBalanceRequest(
       balanceId,
       enrolmentId,
       TaxIdentifier("GB12345678900"),
-      GuaranteeReference("05DE3300BE0001067A001017"),
+      GuaranteeReference("21GB3300BE0001067A001017"),
       Instant.now.minusSeconds(5),
       completedAt = Some(Instant.now),
       Some(balanceRequestSuccess)
     )
 
+    val validResponseXml = Utility
+      .trim(
+        <CD037A>
+          <SynIdeMES1>UNOC</SynIdeMES1>
+          <SynVerNumMES2>3</SynVerNumMES2>
+          <MesSenMES3>NTA.GB</MesSenMES3>
+          <MesRecMES6>MDTP-GUA-00000000000000000000001-01</MesRecMES6>
+          <DatOfPreMES9>20210806</DatOfPreMES9>
+          <TimOfPreMES10>1505</TimOfPreMES10>
+          <IntConRefMES11>{Random.alphanumeric.take(14).mkString}</IntConRefMES11>
+          <MesIdeMES19>{Random.alphanumeric.take(14).mkString}</MesIdeMES19>
+          <MesTypMES20>GB037A</MesTypMES20>
+          <TRAPRIRC1>
+            <TINRC159>GB12345678900</TINRC159>
+          </TRAPRIRC1>
+          <CUSTOFFGUARNT>
+            <RefNumRNT1>GB000001</RefNumRNT1>
+          </CUSTOFFGUARNT>
+          <GUAREF2>
+            <GuaRefNumGRNREF21>21GB3300BE0001067A001017</GuaRefNumGRNREF21>
+            <AccDatREF24>20210114</AccDatREF24>
+            <GuaTypREF22>4</GuaTypREF22>
+            <GuaMonCodREF23>1</GuaMonCodREF23>
+            <GUAQUE>
+              <QueIdeQUE1>2</QueIdeQUE1>
+            </GUAQUE>
+            <EXPEXP>
+              <ExpEXP1>2751.95</ExpEXP1>
+              <ExpCouEXP2>2448</ExpCouEXP2>
+              <BalEXP3>1212211848.45</BalEXP3>
+              <CurEXP4>GBP</CurEXP4>
+            </EXPEXP>
+          </GUAREF2>
+        </CD037A>
+      )
+      .toString
+
     service(
       updateBalanceRequestResponse = IO.some(updatedBalanceRequest)
-    ).updateBalanceRequest(balanceId.messageIdentifier, balanceRequestSuccess)
+    ).updateBalanceRequest(
+      balanceId.messageIdentifier,
+      MessageType.ResponseQueryOnGuarantees,
+      validResponseXml
+    ).map {
+      _ shouldBe Right(updatedBalanceRequest)
+    }.unsafeToFuture()
+  }
+
+  it should "return XML validation error when the XML does not pass schema validation" in {
+    val responseMissingMesTypXml = Utility
+      .trim(
+        <CD037A>
+          <SynIdeMES1>UNOC</SynIdeMES1>
+          <SynVerNumMES2>3</SynVerNumMES2>
+          <MesSenMES3>NTA.GB</MesSenMES3>
+          <MesRecMES6>MDTP-GUA-00000000000000000000001-01</MesRecMES6>
+          <DatOfPreMES9>20210806</DatOfPreMES9>
+          <TimOfPreMES10>1505</TimOfPreMES10>
+          <IntConRefMES11>{Random.alphanumeric.take(14).mkString}</IntConRefMES11>
+          <MesIdeMES19>{Random.alphanumeric.take(14).mkString}</MesIdeMES19>
+          <TRAPRIRC1>
+            <TINRC159>GB12345678900</TINRC159>
+          </TRAPRIRC1>
+          <CUSTOFFGUARNT>
+            <RefNumRNT1>GB000001</RefNumRNT1>
+          </CUSTOFFGUARNT>
+          <GUAREF2>
+            <GuaRefNumGRNREF21>21GB3300BE0001067A001017</GuaRefNumGRNREF21>
+            <AccDatREF24>20210114</AccDatREF24>
+            <GuaTypREF22>4</GuaTypREF22>
+            <GuaMonCodREF23>1</GuaMonCodREF23>
+            <GUAQUE>
+              <QueIdeQUE1>2</QueIdeQUE1>
+            </GUAQUE>
+            <EXPEXP>
+              <ExpEXP1>2751.95</ExpEXP1>
+              <ExpCouEXP2>2448</ExpCouEXP2>
+              <BalEXP3>1212211848.45</BalEXP3>
+              <CurEXP4>GBP</CurEXP4>
+            </EXPEXP>
+          </GUAREF2>
+        </CD037A>
+      )
+      .toString
+
+    service(
+      updateBalanceRequestResponse = IO.none
+    ).updateBalanceRequest(
+      balanceId.messageIdentifier,
+      MessageType.ResponseQueryOnGuarantees,
+      responseMissingMesTypXml
+    ).map {
+      _ shouldBe Left(
+        XmlValidationError(
+          MessageType.ResponseQueryOnGuarantees,
+          NonEmptyList.one(
+            SchemaValidationError(
+              1,
+              332,
+              "cvc-complex-type.2.4.a: Invalid content was found starting with element 'TRAPRIRC1'. One of '{MesTypMES20}' is expected."
+            )
+          )
+        )
+      )
+    }.unsafeToFuture()
+  }
+
+  it should "return bad request error if XML with missing data somehow passes schema validation" in {
+    val responseMissingDataXml =
+      <CD037A>
+        <SynIdeMES1>UNOC</SynIdeMES1>
+        <SynVerNumMES2>3</SynVerNumMES2>
+        <MesSenMES3>NTA.GB</MesSenMES3>
+        <MesRecMES6>MDTP-GUA-00000000000000000000001-01</MesRecMES6>
+        <DatOfPreMES9>20210806</DatOfPreMES9>
+        <TimOfPreMES10>1505</TimOfPreMES10>
+        <IntConRefMES11>{Random.alphanumeric.take(14).mkString}</IntConRefMES11>
+        <MesIdeMES19>{Random.alphanumeric.take(14).mkString}</MesIdeMES19>
+        <MesTypMES20>GB037A</MesTypMES20>
+        <TRAPRIRC1>
+          <TINRC159>GB12345678900</TINRC159>
+        </TRAPRIRC1>
+        <CUSTOFFGUARNT>
+          <RefNumRNT1>GB000001</RefNumRNT1>
+        </CUSTOFFGUARNT>
+        <GUAREF2>
+          <GuaRefNumGRNREF21>21GB3300BE0001067A001017</GuaRefNumGRNREF21>
+          <AccDatREF24>20210114</AccDatREF24>
+          <GuaTypREF22>4</GuaTypREF22>
+          <GuaMonCodREF23>1</GuaMonCodREF23>
+          <GUAQUE>
+            <QueIdeQUE1>2</QueIdeQUE1>
+          </GUAQUE>
+          <EXPEXP>
+            <ExpEXP1>2751.95</ExpEXP1>
+            <ExpCouEXP2>2448</ExpCouEXP2>
+            <CurEXP4>GBP</CurEXP4>
+          </EXPEXP>
+        </GUAREF2>
+      </CD037A>
+
+    val dummyXmlValidator = new XmlValidationService {
+      override def validate(
+        messageType: MessageType,
+        xml: String
+      ): Either[NonEmptyList[SchemaValidationError], Elem] =
+        Right(responseMissingDataXml)
+    }
+
+    service(
+      updateBalanceRequestResponse = IO.none,
+      validator = dummyXmlValidator
+    ).updateBalanceRequest(
+      balanceId.messageIdentifier,
+      MessageType.ResponseQueryOnGuarantees,
+      responseMissingDataXml.toString
+    ).map {
+      _ shouldBe Left(BadRequestError("Unable to parse required values from IE037 message"))
+    }.unsafeToFuture()
+  }
+
+  it should "return not found error if the balance request to update can't be found" in {
+
+    val balanceRequestSuccess =
+      BalanceRequestSuccess(BigDecimal("12345678.90"), CurrencyCode("GBP"))
+
+    val responseMissingDataXml =
+      <CD037A>
+        <SynIdeMES1>UNOC</SynIdeMES1>
+        <SynVerNumMES2>3</SynVerNumMES2>
+        <MesSenMES3>NTA.GB</MesSenMES3>
+        <MesRecMES6>MDTP-GUA-00000000000000000000001-01</MesRecMES6>
+        <DatOfPreMES9>20210806</DatOfPreMES9>
+        <TimOfPreMES10>1505</TimOfPreMES10>
+        <IntConRefMES11>{Random.alphanumeric.take(14).mkString}</IntConRefMES11>
+        <MesIdeMES19>{Random.alphanumeric.take(14).mkString}</MesIdeMES19>
+        <MesTypMES20>GB037A</MesTypMES20>
+        <TRAPRIRC1>
+          <TINRC159>GB12345678900</TINRC159>
+        </TRAPRIRC1>
+        <CUSTOFFGUARNT>
+          <RefNumRNT1>GB000001</RefNumRNT1>
+        </CUSTOFFGUARNT>
+        <GUAREF2>
+          <GuaRefNumGRNREF21>21GB3300BE0001067A001017</GuaRefNumGRNREF21>
+          <AccDatREF24>20210114</AccDatREF24>
+          <GuaTypREF22>4</GuaTypREF22>
+          <GuaMonCodREF23>1</GuaMonCodREF23>
+          <GUAQUE>
+            <QueIdeQUE1>2</QueIdeQUE1>
+          </GUAQUE>
+          <EXPEXP>
+            <ExpEXP1>2751.95</ExpEXP1>
+            <ExpCouEXP2>2448</ExpCouEXP2>
+            <CurEXP4>GBP</CurEXP4>
+          </EXPEXP>
+        </GUAREF2>
+      </CD037A>
+
+    val dummyXmlValidator = new XmlValidationService {
+      override def validate(
+        messageType: MessageType,
+        xml: String
+      ): Either[NonEmptyList[SchemaValidationError], Elem] =
+        Right(responseMissingDataXml)
+    }
+
+    val dummyXmlParser = new XmlParsingService {
+      override def parseResponseMessage(
+        messageType: MessageType,
+        message: NodeSeq
+      ): Either[BalanceRequestError, BalanceRequestResponse] =
+        Right(balanceRequestSuccess)
+    }
+
+    service(
+      updateBalanceRequestResponse = IO.none,
+      validator = dummyXmlValidator,
+      parser = dummyXmlParser
+    )
+      .updateBalanceRequest(
+        balanceId.messageIdentifier,
+        MessageType.ResponseQueryOnGuarantees,
+        responseMissingDataXml.toString
+      )
       .map {
-        _ shouldBe Some(updatedBalanceRequest)
+        _ shouldBe Left(
+          NotFoundError(
+            "The balance request with message identifier MDTP-GUA-22b9899e24ee48e6a18997d1 was not found"
+          )
+        )
       }
       .unsafeToFuture()
   }
